Restaurant_Management_System/Restaurant_Management_System/requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.2.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.1
Werkzeug==2.2.2
gunicorn==20.1.0
requests==2.28.1 
pandas==1.5.1
flask-sqlalchemy==3.0.2
Flask-MySQLdb==1.0.1
mysql-connector-python==8.0.32 
pymysql==1.0.3